+++ cmd/template/cluster/flag.go
@@ -19,13 +19,14 @@ const (
 	flagPodsCIDR     = "pods-cidr"
 
 	// Common.
-	flagClusterID = "cluster-id"
-	flagMasterAZ  = "master-az"
-	flagName      = "name"
-	flagOutput    = "output"
-	flagOwner     = "owner"
-	flagRelease   = "release"
-	flagLabel     = "label"
+	flagClusterID      = "cluster-id"
+	flagControlPlaneAZ = "control-plane-az"
+	flagMasterAZ       = "master-az" // TODO: Remove some time after August 2020
+	flagName           = "name"
+	flagOutput         = "output"
+	flagOwner          = "owner"
+	flagRelease        = "release"
+	flagLabel          = "label"
 )
 
 type flag struct {
@@ -36,13 +37,14 @@ type flag struct {
 	PodsCIDR     string
 
 	// Common.
-	ClusterID string
-	MasterAZ  []string
-	Name      string
-	Output    string
-	Owner     string
-	Release   string
-	Label     []string
+	ClusterID      string
+	ControlPlaneAZ []string
+	MasterAZ       []string
+	Name           string
+	Output         string
+	Owner          string
+	Release        string
+	Label          []string
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
@@ -54,16 +56,25 @@ func (f *flag) Init(cmd *cobra.Command) {
 
 	// Common.
 	cmd.Flags().StringVar(&f.ClusterID, flagClusterID, "", "User-defined cluster ID.")
-	cmd.Flags().StringSliceVar(&f.MasterAZ, flagMasterAZ, []string{}, "Workload master node availability zone.")
+	cmd.Flags().StringSliceVar(&f.ControlPlaneAZ, flagControlPlaneAZ, []string{}, "Availability zone(s) to use by control plane nodes.")
+	cmd.Flags().StringSliceVar(&f.MasterAZ, flagMasterAZ, []string{}, "Replaced by --control-plane-az.")
 	cmd.Flags().StringVar(&f.Name, flagName, "", "Workload cluster name.")
 	cmd.Flags().StringVar(&f.Output, flagOutput, "", "File path for storing CRs.")
 	cmd.Flags().StringVar(&f.Owner, flagOwner, "", "Workload cluster owner organization.")
 	cmd.Flags().StringVar(&f.Release, flagRelease, "", "Workload cluster release. If not given, this remains empty for defaulting to the most recent one via the Management API.")
 	cmd.Flags().StringSliceVar(&f.Label, flagLabel, nil, "Workload cluster label.")
+
+	// TODO: Remove the flag completely some time after August 2020
+	_ = cmd.Flags().MarkDeprecated(flagMasterAZ, "please use --control-plane-az.")
 }
 
 func (f *flag) Validate() error {
 	var err error
+
+	// TODO: Remove the flag completely some time after August 2020
+	if len(f.MasterAZ) > 0 && len(f.ControlPlaneAZ) > 0 {
+		return microerror.Maskf(invalidFlagError, "--control-plane-az and --master-az cannot be combined")
+	}
 
 	if f.Provider != key.ProviderAWS && f.Provider != key.ProviderAzure {
 		return microerror.Maskf(invalidFlagError, "--%s must be either aws or azure", flagProvider)
@@ -110,12 +121,12 @@ func (f *flag) Validate() error {
 		// Validate Master AZs.
 		switch f.Provider {
 		case key.ProviderAWS:
-			if len(f.MasterAZ) != 0 && len(f.MasterAZ) != 1 && len(f.MasterAZ) != 3 {
-				return microerror.Maskf(invalidFlagError, "--%s must be set to either one or three availability zone names", flagMasterAZ)
+			if len(f.ControlPlaneAZ) != 0 && len(f.ControlPlaneAZ) != 1 && len(f.ControlPlaneAZ) != 3 {
+				return microerror.Maskf(invalidFlagError, "--%s must be set to either one or three availability zone names", flagControlPlaneAZ)
 			}
 		case key.ProviderAzure:
-			if len(f.MasterAZ) > 1 {
-				return microerror.Maskf(invalidFlagError, "--%s supports one availability zone only", flagMasterAZ)
+			if len(f.ControlPlaneAZ) > 1 {
+				return microerror.Maskf(invalidFlagError, "--%s supports one availability zone only", flagControlPlaneAZ)
 			}
 		}
 	}