+++ cmd/template/cluster/provider/common.go
@@ -16,7 +16,7 @@ type ClusterCRsConfig struct {
 	// Common.
 	FileName       string
 	ClusterID      string
-	MasterAZ       []string
+	ControlPlaneAZ []string
 	Description    string
 	Owner          string
 	ReleaseVersion string