+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
+- Add the AWS subnet annotations into the `template` command
 - Limit the time allowed for the `login` command to call the authentication proxy to one minute.
 
 ## [1.28.0] - 2021-05-11